+++ gr-digital/lib/additive_scrambler_bb_impl.cc
@@ -113,9 +113,9 @@ namespace gr {
     }
 
     int
-    additive_scrambler_bb_impl::work(int noutput_items,
-				     gr_vector_const_void_star &input_items,
-				     gr_vector_void_star &output_items)
+    additive_scrambler_bb_impl::work(size_t noutput_items,
+                     gr_vector_const_void_star &input_items,
+                     gr_vector_void_star &output_items)
     {
       const unsigned char *in = (const unsigned char *)input_items[0];
       unsigned char *out = (unsigned char *)output_items[0];
